+++ tests/test_client_cursor.py
@@ -2,34 +2,51 @@ import gc
 import pickle
 import weakref
 import datetime as dt
-from typing import List, Union
+from typing import List
 
 import pytest
 
 import psycopg
-from psycopg import pq, sql, rows
+from psycopg import sql, rows
 from psycopg.adapt import PyFormat
 from psycopg.postgres import types as builtins
-from psycopg.rows import RowMaker
 
 from .utils import gc_collect
+from .test_cursor import my_row_factory
+
+
+@pytest.fixture
+def conn(conn):
+    conn.cursor_factory = psycopg.ClientCursor
+    return conn
 
 
 def test_init(conn):
-    cur = psycopg.Cursor(conn)
+    cur = psycopg.ClientCursor(conn)
     cur.execute("select 1")
     assert cur.fetchone() == (1,)
 
     conn.row_factory = rows.dict_row
-    cur = psycopg.Cursor(conn)
+    cur = psycopg.ClientCursor(conn)
     cur.execute("select 1 as a")
     assert cur.fetchone() == {"a": 1}
 
 
 def test_init_factory(conn):
-    cur = psycopg.Cursor(conn, row_factory=rows.dict_row)
+    cur = psycopg.ClientCursor(conn, row_factory=rows.dict_row)
     cur.execute("select 1 as a")
     assert cur.fetchone() == {"a": 1}
+
+
+def test_from_cursor_factory(dsn):
+    with psycopg.connect(dsn, cursor_factory=psycopg.ClientCursor) as aconn:
+        cur = aconn.cursor()
+        assert type(cur) is psycopg.ClientCursor
+
+        cur.execute("select %s", (1,))
+        assert cur.fetchone() == (1,)
+        assert cur._query
+        assert cur._query.query == b"select 1"
 
 
 def test_close(conn):
@@ -131,11 +148,17 @@ def test_statusmessage(conn):
     assert cur.statusmessage is None
 
 
+def test_execute_sql(conn):
+    cur = conn.cursor()
+    cur.execute(sql.SQL("select {value}").format(value="hello"))
+    assert cur.fetchone() == ("hello",)
+
+
 def test_execute_many_results(conn):
     cur = conn.cursor()
     assert cur.nextset() is None
 
-    rv = cur.execute("select 'foo'; select generate_series(1,3)")
+    rv = cur.execute("select %s; select generate_series(1,%s)", ("foo", 3))
     assert rv is cur
     assert cur.fetchall() == [("foo",)]
     assert cur.rowcount == 1
@@ -209,19 +232,15 @@ def test_fetchone(conn):
 
 
 def test_binary_cursor_execute(conn):
-    cur = conn.cursor(binary=True)
-    cur.execute("select %s, %s", [1, None])
-    assert cur.fetchone() == (1, None)
-    assert cur.pgresult.fformat(0) == 1
-    assert cur.pgresult.get_value(0, 0) == b"\x00\x01"
+    with pytest.raises(psycopg.NotSupportedError):
+        cur = conn.cursor(binary=True)
+        cur.execute("select %s, %s", [1, None])
 
 
 def test_execute_binary(conn):
-    cur = conn.cursor()
-    cur.execute("select %s, %s", [1, None], binary=True)
-    assert cur.fetchone() == (1, None)
-    assert cur.pgresult.fformat(0) == 1
-    assert cur.pgresult.get_value(0, 0) == b"\x00\x01"
+    with pytest.raises(psycopg.NotSupportedError):
+        cur = conn.cursor()
+        cur.execute("select %s, %s", [1, None], binary=True)
 
 
 def test_binary_cursor_text_override(conn):
@@ -357,7 +376,10 @@ def test_executemany_rowcount_no_hit(conn, execmany):
     "query",
     [
         "insert into nosuchtable values (%s, %s)",
-        "copy (select %s, %s) to stdout",
+        # This fails, but only because we try to copy in pipeline mode,
+        # crashing the connection. Which would be even fine, but with
+        # the async cursor it's worse... See test_client_cursor_async.py.
+        # "copy (select %s, %s) to stdout",
         "wat (%s, %s)",
     ],
 )
@@ -540,8 +562,8 @@ def test_query_params_execute(conn):
 
     cur.execute("select %t, %s::text", [1, None])
     assert cur._query is not None
-    assert cur._query.query == b"select $1, $2::text"
-    assert cur._query.params == [b"1", None]
+    assert cur._query.query == b"select 1, NULL::text"
+    assert cur._query.params == (b"1", b"NULL")
 
     cur.execute("select 1")
     assert cur._query.query == b"select 1"
@@ -550,16 +572,28 @@ def test_query_params_execute(conn):
     with pytest.raises(psycopg.DataError):
         cur.execute("select %t::int", ["wat"])
 
-    assert cur._query.query == b"select $1::int"
-    assert cur._query.params == [b"wat"]
+    assert cur._query.query == b"select 'wat'::int"
+    assert cur._query.params == (b"'wat'",)
 
 
 def test_query_params_executemany(conn):
     cur = conn.cursor()
 
     cur.executemany("select %t, %t", [[1, 2], [3, 4]])
-    assert cur._query.query == b"select $1, $2"
-    assert cur._query.params == [b"3", b"4"]
+    assert cur._query.query == b"select 3, 4"
+    assert cur._query.params == (b"3", b"4")
+
+
+@pytest.mark.parametrize("ph, params", [("%s", (10,)), ("%(n)s", {"n": 10})])
+def test_copy_out_param(conn, ph, params):
+    cur = conn.cursor()
+    with cur.copy(
+        f"copy (select * from generate_series(1, {ph})) to stdout", params
+    ) as copy:
+        copy.set_types(["int4"])
+        assert list(copy.rows()) == [(i + 1,) for i in range(10)]
+
+    assert conn.info.transaction_status == conn.TransactionStatus.INTRANS
 
 
 def test_stream(conn):
@@ -572,87 +606,6 @@ def test_stream(conn):
         recs.append(rec)
 
     assert recs == [(1, dt.date(2021, 1, 2)), (2, dt.date(2021, 1, 3))]
-
-
-def test_stream_sql(conn):
-    cur = conn.cursor()
-    recs = list(
-        cur.stream(
-            sql.SQL(
-                "select i, '2021-01-01'::date + i from generate_series(1, {}) as i"
-            ).format(2)
-        )
-    )
-
-    assert recs == [(1, dt.date(2021, 1, 2)), (2, dt.date(2021, 1, 3))]
-
-
-def test_stream_row_factory(conn):
-    cur = conn.cursor(row_factory=rows.dict_row)
-    it = iter(cur.stream("select generate_series(1,2) as a"))
-    assert next(it)["a"] == 1
-    cur.row_factory = rows.namedtuple_row
-    assert next(it).a == 2
-
-
-def test_stream_no_row(conn):
-    cur = conn.cursor()
-    recs = list(cur.stream("select generate_series(2,1) as a"))
-    assert recs == []
-
-
-def test_stream_no_col(conn):
-    cur = conn.cursor()
-    recs = list(cur.stream("select"))
-    assert recs == [()]
-
-
-@pytest.mark.parametrize(
-    "query",
-    [
-        "create table test_stream_badq ()",
-        "copy (select 1) to stdout",
-        "wat?",
-    ],
-)
-def test_stream_badquery(conn, query):
-    cur = conn.cursor()
-    with pytest.raises(psycopg.ProgrammingError):
-        for rec in cur.stream(query):
-            pass
-
-
-def test_stream_binary_cursor(conn):
-    cur = conn.cursor(binary=True)
-    recs = []
-    for rec in cur.stream("select generate_series(1, 2)"):
-        recs.append(rec)
-        assert cur.pgresult.fformat(0) == 1
-        assert cur.pgresult.get_value(0, 0) == bytes([0, 0, 0, rec[0]])
-
-    assert recs == [(1,), (2,)]
-
-
-def test_stream_execute_binary(conn):
-    cur = conn.cursor()
-    recs = []
-    for rec in cur.stream("select generate_series(1, 2)", binary=True):
-        recs.append(rec)
-        assert cur.pgresult.fformat(0) == 1
-        assert cur.pgresult.get_value(0, 0) == bytes([0, 0, 0, rec[0]])
-
-    assert recs == [(1,), (2,)]
-
-
-def test_stream_binary_cursor_text_override(conn):
-    cur = conn.cursor(binary=True)
-    recs = []
-    for rec in cur.stream("select generate_series(1, 2)", binary=False):
-        recs.append(rec)
-        assert cur.pgresult.fformat(0) == 0
-        assert cur.pgresult.get_value(0, 0) == str(rec[0]).encode()
-
-    assert recs == [(1,), (2,)]
 
 
 class TestColumn:
@@ -778,7 +731,7 @@ class TestColumn:
 
 def test_str(conn):
     cur = conn.cursor()
-    assert "psycopg.Cursor" in str(cur)
+    assert "psycopg.ClientCursor" in str(cur)
     assert "[IDLE]" in str(cur)
     assert "[closed]" not in str(cur)
     assert "[no result]" in str(cur)
@@ -793,19 +746,16 @@ def test_str(conn):
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize("fmt", PyFormat)
-@pytest.mark.parametrize("fmt_out", pq.Format)
 @pytest.mark.parametrize("fetch", ["one", "many", "all", "iter"])
 @pytest.mark.parametrize("row_factory", ["tuple_row", "dict_row", "namedtuple_row"])
-def test_leak(dsn, faker, fmt, fmt_out, fetch, row_factory):
-    faker.format = fmt
+def test_leak(dsn, faker, fetch, row_factory):
     faker.choose_schema(ncols=5)
     faker.make_records(10)
     row_factory = getattr(rows, row_factory)
 
     def work():
         with psycopg.connect(dsn) as conn:
-            with conn.cursor(binary=fmt_out, row_factory=row_factory) as cur:
+            with psycopg.ClientCursor(conn, row_factory=row_factory) as cur:
                 cur.execute(faker.drop_stmt)
                 cur.execute(faker.create_stmt)
                 with faker.find_insert_problem(conn):
@@ -838,15 +788,22 @@ def test_leak(dsn, faker, fmt, fmt_out, fetch, row_factory):
     assert n[0] == n[1] == n[2], f"objects leaked: {n[1] - n[0]}, {n[2] - n[1]}"
 
 
-def my_row_factory(
-    cursor: Union[psycopg.Cursor[List[str]], psycopg.AsyncCursor[List[str]]]
-) -> RowMaker[List[str]]:
-    if cursor.description is not None:
-        titles = [c.name for c in cursor.description]
+def test_mogrify(conn):
+    cur = conn.cursor()
+    q = cur.mogrify("select 'hello'")
+    assert q == "select 'hello'"
 
-        def mkrow(values):
-            return [f"{value.upper()}{title}" for title, value in zip(titles, values)]
+    q = cur.mogrify("select %s, %s", [1, dt.date(2020, 1, 1)])
+    assert q == "select 1, '2020-01-01'::date"
 
-        return mkrow
-    else:
-        return rows.no_result
+    conn.execute("set client_encoding to utf8")
+    q = cur.mogrify("select %(s)s", {"s": "\u20ac"})
+    assert q == "select '\u20ac'"
+
+    conn.execute("set client_encoding to latin9")
+    q = cur.mogrify("select %(s)s", {"s": "\u20ac"})
+    assert q == "select '\u20ac'"
+
+    conn.execute("set client_encoding to latin1")
+    with pytest.raises(UnicodeEncodeError):
+        cur.mogrify("select %(s)s", {"s": "\u20ac"})